JBIDE-25386: Consolidate tests from 'org.hibernate.eclipse.console.test' into new test plugin 'org.jboss.tools.hibernate.orm.test' - Move 'org.hibernate.eclipse.console.test.utils.tests.DriverDeleteTest' to 'org.jboss.tools.hibernate.orm.test.DriverDeleteTest'
